# spotify/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "spotify.holewinski.dev",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Get health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/current": {
      "get": {
        "summary": "Get current song",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "is_listening": {
                      "type": "boolean"
                    },
                    "ok": {
                      "type": "boolean"
                    },
                    "progress": {
                      "type": "integer"
                    },
                    "track": {
                      "type": "object",
                      "properties": {
                        "artists": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "external_urls": {
                                "type": "object",
                                "properties": {
                                  "spotify": {
                                    "type": "string"
                                  }
                                }
                              },
                              "href": {
                                "type": "string"
                              },
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "duration": {
                          "type": "integer"
                        },
                        "images": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "height": {
                                "type": "integer"
                              },
                              "url": {
                                "type": "string"
                              },
                              "width": {
                                "type": "integer"
                              }
                            }
                          }
                        },
                        "title": {
                          "type": "string"
                        },
                        "track_id": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/callback": {
      "get": {
        "summary": "Get callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "err": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "err": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth": {
      "get": {
        "summary": "Get Spotify authentication",
        "parameters": [
          {
            "name": "passkey",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Temporary Redirect"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "string"
                    },
                    "err": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}